Get rid of the query menu
Day-to-day usage has shown that this is an anti-feature: it takes too
much time to select the query in the list, and it doesn't give incentive
to learn the digit/type-of-query mapping.
This greatly simplifies the code.
Also use consistent variable names.
Also make the menu items safer by using '§' as delimiter.